refactor: 레프리의 역할 분리하기
- WaitingRoom : 입장한 플레이어 기록 및 게임 시작(Referee 생성)
- Referee : 라운드 관리와 점수 및 승패 관리
- PongBoard : 공의 움직임과 벽 및 패들과의 충돌 처리
